Conflicts:
	be/ut/helper.c
	be/ut/main.c

+++ stob/ut/ad.c
@@ -23,19 +23,22 @@
 #include <sys/stat.h>  /* mkdir */
 #include <sys/types.h> /* mkdir */
 
-#include "dtm/dtm.h"     /* m0_dtx */
 #include "lib/arith.h"   /* min64u */
 #include "lib/misc.h"    /* M0_SET0 */
 #include "lib/memory.h"
 #include "lib/errno.h"
 #include "lib/ub.h"
-#include "ut/ut.h"
 #include "lib/assert.h"
+#include "ut/ut.h"
+#include "ut/ast_thread.h"
 
+#include "dtm/dtm.h"     /* m0_dtx */
 #include "stob/stob.h"
 #include "stob/linux.h"
 #include "stob/ad.h"
 #include "balloc/balloc.h"
+
+#include "be/ut/helper.h"
 
 /**
    @addtogroup stob
@@ -79,7 +82,10 @@ static char *read_bufs[NR];
 static m0_bindex_t stob_vec[NR];
 static struct m0_clink clink;
 static struct m0_dtx tx;
-static struct m0_dbenv db;
+struct m0_be_ut_backend	 ut_be;
+struct m0_be_ut_seg	 ut_seg;
+static struct m0_be_seg *db;
+static struct m0_sm_group *sm_grp;
 static uint32_t block_shift;
 static uint32_t buf_size;
 
@@ -93,7 +99,8 @@ static struct mock_balloc *b2mock(struct m0_ad_balloc *ballroom)
 	return container_of(ballroom, struct mock_balloc, mb_ballroom);
 }
 
-static int mock_balloc_init(struct m0_ad_balloc *ballroom, struct m0_dbenv *db,
+static int mock_balloc_init(struct m0_ad_balloc *ballroom, struct m0_be_seg *db,
+			    struct m0_sm_group *grp,
 			    uint32_t bshift, m0_bindex_t container_size,
 			    m0_bcount_t groupsize, m0_bcount_t res_groups)
 {
@@ -142,10 +149,17 @@ struct mock_balloc mb = {
 	}
 };
 
+extern struct m0_sm_group ut__txs_sm_group;
+
 static int test_ad_init(void)
 {
+	struct m0_be_tx_credit cred = {0};
 	int	i;
 	int	result;
+
+	result = m0_ut_ast_thread_start(&ut__txs_sm_group);
+	if (result != 0)
+		goto out;
 
 	result = system("rm -fr ./__s");
 	M0_ASSERT(result == 0);
@@ -156,8 +170,13 @@ static int test_ad_init(void)
 	result = mkdir("./__s/o", 0700);
 	M0_ASSERT(result == 0 || (result == -1 && errno == EEXIST));
 
-	result = m0_dbenv_init(&db, db_name, 0);
-	M0_ASSERT(result == 0);
+	/* Init BE */
+	m0_be_ut_backend_init(&ut_be);
+	m0_be_ut_seg_init(&ut_seg, &ut_be, 1ULL << 24);
+	m0_be_ut_seg_allocator_init(&ut_seg, &ut_be);
+	db = &ut_seg.bus_seg;
+
+	sm_grp = m0_be_ut_backend_sm_group_lookup(&ut_be);
 
 	result = m0_linux_stob_domain_locate("./__s", &dom_back);
 	M0_ASSERT(result == 0);
@@ -170,10 +189,12 @@ static int test_ad_init(void)
 	M0_ASSERT(result == 0);
 	M0_ASSERT(obj_back->so_state == CSS_EXISTS);
 
-	result = m0_ad_stob_domain_locate("", &dom_fore, obj_back);
+	result = m0_ad_stob_domain_locate("", db, sm_grp,
+					  &dom_fore, obj_back);
 	M0_ASSERT(result == 0);
 
-	result = m0_ad_stob_setup(dom_fore, &db, obj_back, &mb.mb_ballroom,
+	result = m0_ad_stob_setup(dom_fore, db, sm_grp, obj_back,
+				  &mb.mb_ballroom,
 				  BALLOC_DEF_CONTAINER_SIZE,
 				  BALLOC_DEF_BLOCK_SHIFT,
 				  BALLOC_DEF_BLOCKS_PER_GROUP,
@@ -186,7 +207,18 @@ static int test_ad_init(void)
 	M0_ASSERT(result == 0);
 	M0_ASSERT(obj_fore->so_state == CSS_UNKNOWN);
 
-	m0_dtx_init(&tx);
+	block_shift = obj_fore->so_op->sop_block_shift(obj_fore);
+	/* buf_size is chosen so it would be at least MIN_BUF_SIZE in bytes
+	 * or it would consist of at least MIN_BUF_SIZE_IN_BLOCKS blocks */
+	buf_size = max_check(MIN_BUF_SIZE
+			, (1 << block_shift) * MIN_BUF_SIZE_IN_BLOCKS);
+
+	m0_dtx_init(&tx, db->bs_domain, sm_grp);
+	dom_fore->sd_ops->sdo_write_credit(dom_fore,
+				buf_size * (NR * NR / 2) /* test_ad() */ +
+				buf_size * NR, /* test_ad_rw_unordered() */
+						&cred);
+	m0_be_tx_prep(&tx.tx_betx, &cred);
 	result = dom_fore->sd_ops->sdo_tx_make(dom_fore, &tx);
 	M0_ASSERT(result == 0);
 
@@ -197,12 +229,6 @@ static int test_ad_init(void)
 		M0_ASSERT(result == 0);
 	}
 	M0_ASSERT(obj_fore->so_state == CSS_EXISTS);
-
-	block_shift = obj_fore->so_op->sop_block_shift(obj_fore);
-	/* buf_size is chosen so it would be at least MIN_BUF_SIZE in bytes
-	 * or it would consist of at least MIN_BUF_SIZE_IN_BLOCKS blocks */
-	buf_size = max_check(MIN_BUF_SIZE
-			, (1 << block_shift) * MIN_BUF_SIZE_IN_BLOCKS);
 
 	for (i = 0; i < ARRAY_SIZE(user_buf); ++i) {
 		user_buf[i] = m0_alloc_aligned(buf_size, block_shift);
@@ -221,6 +247,7 @@ static int test_ad_init(void)
 		stob_vec[i] = (buf_size * (2 * i + 1)) >> block_shift;
 		memset(user_buf[i], ('a' + i)|1, buf_size);
 	}
+out:
 	return result;
 }
 
@@ -231,15 +258,21 @@ static int test_ad_fini(void)
 	m0_dtx_done(&tx);
 
 	m0_stob_put(obj_fore);
-	dom_fore->sd_ops->sdo_fini(dom_fore);
-	dom_back->sd_ops->sdo_fini(dom_back);
-	m0_dbenv_fini(&db);
+	dom_fore->sd_ops->sdo_fini(dom_fore, sm_grp);
+	dom_back->sd_ops->sdo_fini(dom_back, NULL);
+
+	m0_be_ut_seg_allocator_fini(&ut_seg, &ut_be);
+	m0_be_ut_seg_fini(&ut_seg);
+	m0_be_ut_backend_fini(&ut_be);
 
 	for (i = 0; i < ARRAY_SIZE(user_buf); ++i)
 		m0_free(user_buf[i]);
 
 	for (i = 0; i < ARRAY_SIZE(read_buf); ++i)
 		m0_free(read_buf[i]);
+
+	m0_ut_ast_thread_stop();
+
 	return 0;
 }
 
@@ -365,12 +398,16 @@ static void test_ad(void)
 
 static void test_ad_undo(void)
 {
+	struct m0_be_tx_credit  cred = {0};
 	int                     result;
 	struct m0_fol_rec_part *rpart;
 
 	m0_dtx_done(&tx);
 
-	m0_dtx_init(&tx);
+	m0_dtx_init(&tx, db->bs_domain, sm_grp);
+	dom_fore->sd_ops->sdo_write_credit(dom_fore,
+				buf_size * 2, &cred);
+	m0_be_tx_prep(&tx.tx_betx, &cred);
 	result = dom_fore->sd_ops->sdo_tx_make(dom_fore, &tx);
 	M0_UT_ASSERT(result == 0);
 
@@ -389,11 +426,11 @@ static void test_ad_undo(void)
 	test_write(1);
 
 	/* Do the undo operation. */
-	result = rpart->rp_ops->rpo_undo(rpart, &tx.tx_dbtx);
+	result = rpart->rp_ops->rpo_undo(rpart, &tx.tx_betx);
 	M0_UT_ASSERT(result == 0);
 	m0_dtx_done(&tx);
 
-	m0_dtx_init(&tx);
+	m0_dtx_init(&tx, db->bs_domain, sm_grp);
 	result = dom_fore->sd_ops->sdo_tx_make(dom_fore, &tx);
 	M0_UT_ASSERT(result == 0);
 	test_read(1);
